requirements.txt
faiss-cpu==1.8.0.post1
sentence-transformers>=2.7.0
pypdf>=4.2.0
gradio>=4.37.2
openai>=1.35.7
requests>=2.32.3
numpy>=1.26.4
python-dotenv